refactor form-select-organization logic
